// src/API.ts
export type Chat = {
  __typename: "Chat",
  id: string,
  chatroomId: string,
  text: string,
  sender: string,
  createdAt: string,
  updatedAt: string,
};